web_service/tg_bot/requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
loguru==0.7.2
peewee==3.17.0
aiohttp==3.8.6
aiohttp-proxy==0.1.2
openpyxl==3.1.2
psutil==5.9.6
openai==0.27.2
psycopg2-binary==2.9.9